HW4/digest/requirements.txt
Flask==3.0.3
waitress==3.0.0
flask_httpauth==4.8.0
python-dotenv==1.0.1